fix: mypy 94→0 errors, all type annotations clean
- Added TYPE_CHECKING imports for forward references in all model files
  (tenant.py, source.py, document.py, chunk.py)
- Fixed dict → dict[str, Any] across all files (agent-applied)
- Fixed Row[Any] | None null checks in structured.py
- Fixed type: ignore comments for CursorResult.rowcount
- Removed duplicate asyncio import in claims.py
- Fixed line length in pipeline.py

Triple-green: ruff 0 errors, pytest 131 passed, mypy 0 errors.

+++ src/raasoa/quality/claim_conflicts.py
@@ -180,4 +180,4 @@ def _cosine_similarity(a: list[float], b: list[float]) -> float:
     norm_b = sum(x * x for x in b) ** 0.5
     if norm_a == 0 or norm_b == 0:
         return 0.0
-    return dot / (norm_a * norm_b)
+    return float(dot / (norm_a * norm_b))
